requirements.txt
openai
langchain
pinecone-client
python-dotenv
tiktoken
streamlit
tabulate
langchain-openai
langchain-experimental